requirements.txt
asgiref==3.2.10
autopep8==1.5.4
certifi==2020.6.20
chardet==3.0.4
defusedxml==0.6.0
Django==3.1
django-allauth==0.42.0
django-taggit==1.3.0
django-template-maths==0.2.0
idna==2.10
oauthlib==3.1.0
Pillow==7.2.0
pycodestyle==2.6.0
python3-openid==3.2.0
pytz==2020.1
requests==2.24.0
requests-oauthlib==1.3.0
sqlparse==0.3.1
toml==0.10.1
urllib3==1.25.10
